Replace receipt printer Store/Action with async Yosemite service
Drop the Flux Store/Action layer in favor of a manufacturer-agnostic
async ReceiptPrinterService in Yosemite, usable from both POS and the
app target. Move the consumer seam protocol into Yosemite, delete the
app-target adaptor and the POS-owned protocol, add a shared PrinterError,
and lock-guard the connected printer in StarPrinterService.
