+++ Modules/Sources/Hardware/Printer/PrinterError.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// Errors surfaced by a printer discovery/connection service, independent of manufacturer.
+public enum PrinterError: Error {
+    case discoveryFailure
+    case printerNotFound
+}
