+++ WooCommerce/Classes/Yosemite/AuthenticatedState.swift
@@ -133,10 +133,6 @@ class AuthenticatedState: StoresManagerState {
                          network: network,
                          receiptPrinterService: ServiceLocator.receiptPrinterService,
                          fileStorage: PListFileStorage()),
-            ReceiptPrinterStore(dispatcher: dispatcher,
-                                storageManager: storageManager,
-                                network: network,
-                                printerDiscoveryService: ServiceLocator.printerDiscoveryService),
             AnnouncementsStore(dispatcher: dispatcher,
                                storageManager: storageManager,
                                network: network,
